+++ src/Schema/Normalizer/GeoCoordinatesNormalizer.php
@@ -38,8 +38,8 @@ class GeoCoordinatesNormalizer extends StructuredValueNormalizer
 
         $structuredValueNormalized = parent::normalize($registry, $entity, $usage, $context, true);
         $normalized = [
-            "address" => $registry->normalize($entity->getAddress(), $usage, $context, true),
-            "addressCountry" => $registry->normalize($entity->getAddressCountry(), $usage, $context, true),
+            "address" => $this->normalizeDataOrPrimitive($registry, $entity->getAddress(), $usage, $context, true),
+            "addressCountry" => $this->normalizeDataOrPrimitive($registry, $entity->getAddressCountry(), $usage, $context, true),
             "elevation" => $entity->getElevation(),
             "latitude" => $entity->getLatitude(),
             "longitude" => $entity->getLongitude(),